♻️ extract Router Class from index.ts
